+++ src/constants/index.ts
@@ -1,2 +1,2 @@
 export { navigationLinks, socialLinks } from "./links";
-export { emailComponents, emailTemplates } from "./emails";
+export { emailTemplates } from "./emails";
