--- src/email-components/dividers/Sample.preview.tsx
@@ -1,10 +1,0 @@
-import { Html } from '@react-email/html';
-import { Email } from './Sample';
-
-export default function Preview() {
-  return (
-    <Html lang="en" dir="ltr">
-      <Email />
-    </Html>
-  );
-};
