--- src/email-components/dividers/Sample.tsx
@@ -1,5 +1,0 @@
-import { Text } from '@react-email/text';
-
-export const Email = () => {
-  return <Text>Lorem ipsum</Text>;
-};
