+++ src/examples/components/emojis/Sample.tsx
@@ -1,0 +1,10 @@
+import { Html } from '@react-email/html';
+import { Text } from '@mailingui/components';
+
+export default function Preview() {
+  return (
+    <Html lang="en" dir="ltr">
+      <Text />
+    </Html>
+  );
+};
