+++ src/examples/components/heroes/SimpleHero.tsx
@@ -1,0 +1,24 @@
+import * as React from "react";
+import { Body, Container, Head, Html, Preview } from "@react-email/components";
+import { Heading, theme } from "@mailingui/components";
+
+const SimpleHero = () => (
+  <Html>
+    <Head />
+    <Preview>Simple Hero</Preview>
+    <Body style={main}>
+      <Container style={theme.container}>
+        <Heading />
+      </Container>
+    </Body>
+  </Html>
+);
+
+export default SimpleHero;
+
+const main = {
+  backgroundColor: theme.backgroundColor,
+  color: "#24292e",
+  fontFamily:
+    '-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"',
+};
