+++ src/mailingui/components/index.ts
@@ -1,0 +1,4 @@
+export { Badge, type BadgeProps } from "./badges/Badges"
+export { Heading, type HeadingProps } from './headings/Heading'
+export { Text } from './texts/Text'
+export { theme } from './theme'
