+++ src/mailingui/components/texts/Text.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { Text as ReactEmailText } from '@react-email/text';
+
+const Text: React.FC = () => {
+  return <ReactEmailText>Lorem ipsum</ReactEmailText>;
+};
+
+export { Text }
